requirements.txt
requests==2.26.0
pandas==1.3.3
tensorflow==2.6.0
numpy==1.19.5
scikit-learn==0.24.2
PyYAML==5.4.1
python-dotenv==0.19.1